Breaking up events logic for easier testing
