+++ pyproject.toml
@@ -18,6 +18,7 @@ ddtrace = "^2.3.1"
 pip = ">=23.1.0"
 flake8 = "^6.1.0"
 black = "^23.12.0"
+pytest = "~7.4.3"
 
 [tool.black]
 line-length = 120
